requirements.txt
binance-futures-connector
scikit-learn~=1.6.1
matplotlib~=3.10.0
requests~=2.32.3
colorama~=0.4.6
schedule~=1.2.2
xgboost~=2.1.4
joblib~=1.4.2
pandas~=2.2.3
imblearn~=0.0
numpy~=2.2.1
ta~=0.11.0
